requirements.txt
django-cms<3.3
django-treebeard>=2.0
psycopg2
https://github.com/divio/djangocms-column/archive/master.zip?1460184928.2004218
https://github.com/divio/djangocms-googlemap/archive/master.zip?1460184928.2004218
https://github.com/divio/djangocms-inherit/archive/master.zip?1460184928.2004218
https://github.com/divio/djangocms-link/archive/master.zip?1460184928.2004218
https://github.com/divio/djangocms-style/archive/master.zip?1460184928.2004218
easy_thumbnails
django-filer>=0.9.12
cmsplugin-filer>=0.10.2
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.1
Django<1.10
pytz
django-reversion>=1.10,<1.11
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six

django-taggit
django-parler
#mptt
django-extended-choices
django-standard_form
#absolute
#emailit
#captcha
django-bootstrap3
django-admin-sortable2
#sortedm2m
#easy_thumbnails
PyYAML

aldryn_common
aldryn_boilerplates
aldryn_reversion
aldryn_apphooks_config
aldryn_apphook_reload

aldryn-bootstrap3
aldryn-forms
aldryn-style

#aldryn-newsblog
#aldryn-people
#aldryn-events
#aldryn-faq
#aldryn-jobs
#aldryn-locations
#aldryn-gallery
#aldryn-categories

djangorestframework
djangorestframework-recursive
https://github.com/inonit/djangocms-restapi/archive/master.zip
